ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
asyncpg==0.29.0 